# src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Bank Cards API
  description: |
    API для управления банковскими картами и пользователями.
    Предоставляет функционал для регистрации, аутентификации, управления картами,
    перевода средств между картами и администрирования пользователей.
    
    ## Аутентификация
    Для доступа к защищенным endpoints требуется JWT-токен, полученный при логине.
    Добавьте токен в заголовок запроса: `Authorization: Bearer <your_token>`
    
    ## Роли пользователей
    - **ROLE_USER** - обычный пользователь (может управлять своими картами)
    - **ROLE_ADMIN** - администратор (полный доступ ко всем операциям)
  version: 1.0.0
servers:
  - url: http://localhost:8082/api/v1
    description: Development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Bad Request
        errorDescription:
          type: string
          example: Invalid input parameters
        statusCode:
          type: integer
          example: 400
        path:
          type: string
          example: /api/v1/cards
        timestamp:
          type: string
          format: date-time
          example: "2023-12-01T10:30:00"
...
    UserRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
        - password
      properties:
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        password:
          type: string
          format: password
          example: "SecurePassword123!"
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
    PageResponse:
      type: object
      properties:
        content:
          type: array
          items: { }
        currentPage:
          type: integer
          example: 0
        totalElements:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
...
security:
  - BearerAuth: [ ]
paths:
...
  /users:
    post:
      tags: [ Users ]
      summary: Создать нового пользователя (только ADMIN)
      description: |
        Создает нового пользователя. Доступно только администраторам.
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '201':
          description: Пользователь успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [ Users ]
      summary: Получить список всех пользователей (только ADMIN)
      description: |
        Возвращает paginated список всех пользователей в системе.
        Доступно только администраторам.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: currentPage
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Количество элементов на странице (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Поисковый запрос
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Список пользователей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
              examples:
                users-page:
                  summary: Пример страницы пользователей
                  value:
                    content:
                      - id: "123e4567-e89b-12d3-a456-426614174000"
                        firstName: "John"
                        lastName: "Doe"
                        email: "john.doe@example.com"
                      - id: "223e4567-e89b-12d3-a456-426614174000"
                        firstName: "Jane"
                        lastName: "Smith"
                        email: "jane.smith@example.com"
                    currentPage: 0
                    totalElements: 2
                    totalPages: 1
  /users/me:
    get:
      tags: [ Users ]
      summary: Получить детали своего аккаунта
      description: |
        Пользователь получает свои данные.
      security:
        - BearerAuth: [ ]
      responses:
        '200':
          description: Детали пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}:
    get:
      tags: [ Users ]
      summary: Получить детали пользователя (только ADMIN)
      description: |
        Возвращает детальную информацию о пользователе.
        Пользователь может получить только свои данные, администратор - любые.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Детали пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [ Users ]
      summary: Обновить данные пользователя
      description: |
        Обновляет данные пользователя. Пользователь может обновить только свои данные,
        администратор - любые.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRequest'
      responses:
        '200':
          description: Данные пользователя обновлены
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Невалидные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь с таким email уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [ Users ]
      summary: Удалить пользователя (только ADMIN)
      description: |
        Удаляет пользователя из системы. Доступно только администраторам.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь успешно удален
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/activate:
    patch:
      tags: [ Users ]
      summary: Активировать пользователя (только ADMIN)
      description: |
        Активирует деактивированного пользователя. Доступно только администраторам.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь активирован
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/{userId}/deactivate:
    patch:
      tags: [ Users ]
      summary: Деактивировать пользователя
      description: |
        Деактивирует пользователя. Пользователь может деактивировать только себя,
        администратор - любого пользователя.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Пользователь деактивирован
        '404':
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /cards/user/me:
    get:
      tags: [ Cards ]
      summary: Получить карты пользователя
      description: |
        Возвращает paginated список карт текущего пользователя.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: currentPage
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Количество элементов на странице (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Поисковый запрос
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
  /cards/user/{userId}:
    get:
      tags: [ Cards ]
      summary: Получить карты пользователя
      description: |
        Возвращает paginated список карт указанного пользователя.
        Пользователь может получить только свои карты, администратор - любые.
      security:
        - BearerAuth: [ ]
      parameters:
        - name: userId
          in: path
          required: true
          description: UUID пользователя
          schema:
            type: string
            format: uuid
        - name: currentPage
          in: query
          description: Номер страницы (начинается с 0)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: limit
          in: query
          description: Количество элементов на странице (1-100)
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: search
          in: query
          description: Поисковый запрос
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Список карт пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse'
...